requirements.txt
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.24.0
pytest>=7.4.0
jinja2>=3.1.0
loguru>=0.7.0
anthropic>=0.25.0
